Fix errors for anchor in non loaded worlds
When EpicAnchor serialized the location from the configuration file, it requires the world to be loaded. Another plugin loading its (e.g. SkyBlock) world later in the chain, causes our Location to have the world `null` - A full rewrite on that system might be better in the future that listens to a world load event and loads that anchors from the config as required